refactor: create app factory with dependency injection
